validator/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
requests>=2.31.0
httpx>=0.27.0
groq>=0.4.1
python-dotenv>=1.0.0
pytest>=7.4.0
